% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Breastcancer.R
\docType{data}
\name{Breastcancer}
\alias{Breastcancer}
\title{A survival data on breast cancer.}
\format{
An object of class \code{data.frame} with 660 rows and 12 columns.
}
\usage{
data(Breastcancer)
}
\description{
A survival data on breast cancer.
}
\examples{
data(Breastcancer)

}
\keyword{datasets}
