% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.tcdca.R
\name{cox.tcdca}
\alias{cox.tcdca}
\title{cox.tcdca}
\arguments{
\item{...}{Fill in multiple survival analysis or competitive risk models. You cannot mix and match.}

\item{newdata}{If the decision curve of the validation set is to be analysed. Fill in the validation set data here.}

\item{timepoint}{If it is a survival analysis, fill in the point in time you need to study. The default is the median time.}

\item{cmprsk}{If it is a competitive risk model, select TRUE here.}

\item{modelnames}{Defines the name of the generated image model.}

\item{y.min}{The maximum value of the negative part of the picture. Generally defaults to positive values multiplied by 0.4.}

\item{xstop}{The maximum value of the X-axis of the picture.}

\item{y.max}{The maximum value of the Y-axis. The default value is the maximum net benefit.}

\item{pyh}{The height at which the bars are plotted cannot exceed y.min.}

\item{relcol}{The colour of the relevant part of the bar. The default is red.}

\item{irrelcol}{The colour of the irrelevant part of the bar. The default is blue.}

\item{relabel}{Relevance Tags.}

\item{irrellabel}{No relevant tags.}

\item{text.size}{Font size.}

\item{text.col}{The colour of the font.}

\item{colbar}{The default is true, and if false is selected, bar plotting is cancelled.}

\item{merge}{If true is selected it will merge the two long zones.}

\item{threshold.text}{The default is FALSE, if TRUE is selected, a text message for the threshold will be added.}

\item{threshold.line}{The default is FALSE, and if TRUE is selected, lines for the threshold will be added.}

\item{nudge_x}{Used to adjust the x-axis position of the point where the threshold is located.}

\item{nudge_y}{Used to adjust the y-axis position of the point where the threshold is located.}

\item{threshold.linetype}{The line shape of the threshold line.}

\item{threshold.linewidth}{The line width of the threshold line.}

\item{threshold.linecol}{The colour of the threshold line.}

\item{po.text.size}{The size of the threshold point text.}

\item{po.text.col}{The colour of the threshold point text.}

\item{po.text.fill}{The background of the threshold point text.}

\item{liftpec}{Threshold point left displacement.}

\item{rightpec}{Threshold point right displacement.}

\item{legend.position}{Set the position of the legend.}

\item{Splitface}{Name the faceted image.}
}
\value{
A picture.
}
\description{
You can use it to plot decision curves for multiple generative analysis or competitive risk models.
}
