% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\docType{data}
\name{demo}
\alias{demo}
\title{A medical examination related data.}
\format{
An object of class \code{data.frame} with 832 rows and 34 columns.
}
\usage{
data(demo)
}
\description{
A medical examination related data.
}
\examples{
data(demo)

}
\keyword{datasets}
