% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_surv.R
\docType{data}
\name{df_surv}
\alias{df_surv}
\title{A data for competitive risk modelling.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 750 rows and 9 columns.
}
\usage{
data(df_surv)
}
\description{
A data for competitive risk modelling.
}
\examples{
data(df_surv)

}
\keyword{df_surv}
