% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIRI.R
\docType{data}
\name{LIRI}
\alias{LIRI}
\title{A data for random forest analysis.}
\format{
An object of class \code{data.frame} with 232 rows and 6 columns.
}
\usage{
data(LIRI)
}
\description{
A data for random forest analysis.
}
\examples{
data(LIRI)

}
\keyword{LIRI}
