% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newcrr.R
\name{newcrr}
\alias{newcrr}
\title{newcrr}
\usage{
newcrr(fit, cencode = 0, failcode = 1)
}
\arguments{
\item{fit}{Modelling for Survival Analysis.}

\item{cencode}{Censor status, default is 0.}

\item{failcode}{Events of interest, default is 1.}
}
\value{
A list of competing risk model formats.
}
\description{
Types of transformation of survival analysis models into competitive risk models.
}
